refact: move Playback.wait to its own extension file
